# code/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.ErrorDetail:
    properties:
      code:
        example: invalid_input
        type: string
      error:
        example: invalid request body
        type: string
    type: object
  handlers.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/handlers.ErrorDetail'
    type: object
...
  suggestion.ChangeSuggestionInput:
    properties:
      account_code:
        type: string
      category_code:
        type: string
      description_contains:
        type: string
      entry_type:
        type: string
      priority:
        type: integer
      transfer_account_code:
        type: string
    type: object
  suggestion.CreateSuggestionInput:
    properties:
      account_code:
        type: string
      category_code:
        type: string
      description_contains:
        type: string
      entry_type:
        type: string
      priority:
        type: integer
      transfer_account_code:
        type: string
    required:
    - description_contains
    - priority
    type: object
  suggestion.SuggestionEntryType:
    enum:
    - REVENUE
    - EXPENSE
    - TRANSFER
    type: string
    x-enum-varnames:
    - SuggestionEntryTypeRevenue
    - SuggestionEntryTypeExpense
    - SuggestionEntryTypeTransfer
  suggestion.SuggestionListResponse:
    properties:
      suggestions:
        items:
          $ref: '#/definitions/suggestion.SuggestionResponseItem'
        type: array
    type: object
  suggestion.SuggestionResponse:
    properties:
      suggestion:
        $ref: '#/definitions/suggestion.SuggestionResponseItem'
    type: object
  suggestion.SuggestionResponseItem:
    properties:
      account_code:
        type: string
      category_code:
        type: string
      created_at:
        type: string
      description_contains:
        type: string
      entry_type:
        $ref: '#/definitions/suggestion.SuggestionEntryType'
      id:
        type: string
      priority:
        type: integer
      transfer_account_code:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  description: API for managing accounts, categories, transactions, reports, and suggestions.
  title: Expense Tracker API
  version: "1.0"
paths:
...
  /suggestions:
    get:
      description: Returns all suggestions for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/suggestion.SuggestionListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List suggestions
      tags:
      - suggestions
    post:
      consumes:
      - application/json
      description: Creates a transaction suggestion for the authenticated user
      parameters:
      - description: Suggestion payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/suggestion.CreateSuggestionInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/suggestion.SuggestionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create suggestion
      tags:
      - suggestions
  /suggestions/{id}:
    delete:
      description: Deletes a suggestion
      parameters:
      - description: Suggestion ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete suggestion
      tags:
      - suggestions
    get:
      description: Returns a suggestion by its ID
      parameters:
      - description: Suggestion ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/suggestion.SuggestionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get suggestion by ID
      tags:
      - suggestions
    patch:
      consumes:
      - application/json
      description: Updates an existing suggestion
      parameters:
      - description: Suggestion ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Suggestion changes
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/suggestion.ChangeSuggestionInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/suggestion.SuggestionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update suggestion
      tags:
      - suggestions
...
securityDefinitions:
  BearerAuth:
    description: JWT token using the format "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"